Create demo
I wanted a menu that closed one item if another was opened. (With most people on the internet with their phones there is not a lot of space in the viewport.) I also wanted something that closed the menu when clicking elsewhere on the screen. Looking on Stack Overflow, I combined a bunch of suggestions and tweaked them till they worked okay together.